split off front end class
